[I25-154] feat : 기여자 여러명일 시 프로필 접어서 사진으로 나열
